engine/pyproject.toml
[project]
name = "avatar-engine"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=23.0",
    "playwright>=1.40",
    "websockets>=12.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
stt = ["faster-whisper>=1.0", "sounddevice>=0.4"]
tts = ["sounddevice>=0.4"]
ml = ["torch>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.12"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["engine*", "server*", "modules*", "profiles*", "orchestrator*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
